requirements.txt
# FastAPI 相关
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain 相关
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-core==0.1.10

# Gradio 相关
gradio==4.7.1

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# 向量数据库相关 (支持多种向量数据库)
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Elasticsearch 相关
elasticsearch==8.11.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 增强 RAG：Redis 上下文、RAGflow 调用
redis==5.0.1

# 其他工具
httpx==0.25.2
aiofiles==23.2.1

# 测试
pytest==7.4.3
pytest-cov==4.1.0
